analysis/pyproject.toml
[project]
name = "analysis"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9.11",
]
